requirements.txt
Pillow
Django
pytz
mysqlclient
setuptools
django-markdownx
django_markdown2